Use cardonnay, reduce code duplication

+++ cardano_node_tests/prepare_cluster_scripts.py
@@ -10,10 +10,11 @@ import pathlib as pl
 import shutil
 import sys
 
+from cardonnay import local_scripts as cardonnay_local
+
 import cardano_node_tests.utils.types as ttypes
 from cardano_node_tests.utils import cluster_nodes
 from cardano_node_tests.utils import cluster_scripts
-from cardano_node_tests.utils import helpers
 
 LOGGER = logging.getLogger(__name__)
 
@@ -24,14 +25,12 @@ def get_args() -> argparse.Namespace:
     parser.add_argument(
         "-d",
         "--dest-dir",
-        required=True,
         help="Path to destination directory",
     )
     parser.add_argument(
-        "-s",
-        "--scripts-dir",
-        type=helpers.check_dir_arg,
-        help="Path to directory with scripts templates",
+        "-t",
+        "--testnet-variant",
+        help="Testnet variant to use.",
     )
     parser.add_argument(
         "-i",
@@ -39,6 +38,12 @@ def get_args() -> argparse.Namespace:
         type=int,
         default=0,
         help="Instance number in the sequence of cluster instances (default: 0)",
+    )
+    parser.add_argument(
+        "-l",
+        "--ls",
+        action="store_true",
+        help="List available testnet variants and exit.",
     )
     parser.add_argument(
         "-c",
@@ -51,25 +56,26 @@ def get_args() -> argparse.Namespace:
 
 def prepare_scripts_files(
     destdir: ttypes.FileType,
-    scriptsdir: ttypes.FileType = "",
+    testnet_variant: str,
     instance_num: int = 0,
-) -> cluster_scripts.InstanceFiles:
+) -> cardonnay_local.InstanceFiles:
     """Prepare scripts files for starting and stopping cluster instance."""
-    start_script: ttypes.FileType = ""
-    stop_script: ttypes.FileType = ""
+    scriptsdir = cluster_scripts.get_testnet_variant_scriptdir(testnet_variant=testnet_variant)
+    if not scriptsdir:
+        msg = f"Testnet variant '{testnet_variant}' is not supported."
+        raise RuntimeError(msg)
 
-    if scriptsdir:
-        scriptsdir = pl.Path(scriptsdir)
-        start_script = next(scriptsdir.glob("start-cluster*"), "")
-        if not start_script:
-            msg = f"Start script not found in '{scriptsdir}'."
-            raise RuntimeError(msg)
+    scriptsdir = pl.Path(scriptsdir)
+
+    testnet_path = scriptsdir / "testnet.json"
+    if not testnet_path:
+        msg = f"Testnet file not found in '{scriptsdir}'."
+        raise RuntimeError(msg)
 
     startup_files = cluster_nodes.get_cluster_type().cluster_scripts.prepare_scripts_files(
         destdir=destdir,
         instance_num=instance_num,
-        start_script=start_script,
-        stop_script=stop_script,
+        scriptsdir=scriptsdir,
     )
     return startup_files
 
@@ -78,7 +84,18 @@ def main() -> int:
     logging.basicConfig(format="%(levelname)s:%(message)s", level=logging.INFO)
     args = get_args()
 
+    if args.ls:
+        variants_str = "\n".join(f" - {v}" for v in cluster_scripts.get_testnet_variants())
+        LOGGER.info(f"Available testnet variants:\n{variants_str}")
+        return 0
+
+    if not args.dest_dir:
+        LOGGER.error("The 'destdir' must be set.")
     destdir = pl.Path(args.dest_dir)
+
+    testnet_variant = args.testnet_variant
+    if not testnet_variant:
+        LOGGER.error("The testnet variant must be set.")
 
     if args.clean:
         shutil.rmtree(destdir, ignore_errors=True)
@@ -89,11 +106,9 @@ def main() -> int:
 
     destdir.mkdir(parents=True)
 
-    scriptsdir: ttypes.FileType = args.scripts_dir or ""
-
     try:
         prepare_scripts_files(
-            destdir=destdir, scriptsdir=scriptsdir, instance_num=args.instance_num
+            destdir=destdir, testnet_variant=testnet_variant, instance_num=args.instance_num
         )
     except Exception:
         LOGGER.exception("Failure")
